Move package of SplashScreen
