+++ app/src/main/java/com/jik/movie/MainActivity.kt
@@ -3,6 +3,7 @@ package com.jik.movie
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import com.jik.movie.splash.createSplashScreen
 import com.jik.movie.ui.MovieApp
 import dagger.hilt.android.AndroidEntryPoint
 import kotlin.time.Duration.Companion.seconds
