+++ app/src/main/java/com/jik/movie/splash/SplashScreen.kt
@@ -1,4 +1,4 @@
-package com.jik.movie
+package com.jik.movie.splash
 
 import android.app.Activity
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
